Minor code cleanups
