+++ riptide-spring-boot-starter/src/test/java/org/zalando/riptide/spring/AccessTokensMissingTest.java
@@ -5,7 +5,6 @@ import com.google.common.base.Throwables;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-import org.springframework.beans.PropertyBatchUpdateException;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.web.HttpMessageConverters;
 import org.springframework.boot.builder.SpringApplicationBuilder;
@@ -14,8 +13,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 
 import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.hasItemInArray;
-import static org.hamcrest.Matchers.instanceOf;
 import static org.hobsoft.hamcrest.compose.ComposeMatchers.compose;
 import static org.hobsoft.hamcrest.compose.ComposeMatchers.hasFeature;
 import static org.mockito.Mockito.mock;
@@ -44,12 +41,10 @@ public final class AccessTokensMissingTest {
 
     @Test
     public void shouldFailDueToMissingAccessTokenUrl() {
-        exception.expectCause(instanceOf(PropertyBatchUpdateException.class));
-        exception.expectCause(hasFeature("exception", PropertyBatchUpdateException::getPropertyAccessExceptions,
-                hasItemInArray(hasFeature("cause", Throwables::getRootCause,
-                        hasFeature("message", Throwable::getMessage,
-                                compose(containsString("riptide.oauth.access-token-url")).and(
-                                        containsString("ACCESS_TOKEN_URL")))))));
+        exception.expect(hasFeature(Throwables::getRootCause,
+                hasFeature("message", Throwable::getMessage,
+                        compose(containsString("riptide.oauth.access-token-url")).and(
+                                containsString("ACCESS_TOKEN_URL")))));
 
         new SpringApplicationBuilder(TestConfiguration.class)
                 .profiles("default")
